refactor: Update import reference to house directory
