+++ app/src/main/java/com/example/dtthouses/data/exception/NetworkException.kt
@@ -1,6 +1,6 @@
 package com.example.dtthouses.data.exception
 
-import com.example.dtthouses.data.contstant.Error.EXCEPTION_ERROR_PREFIX
+import com.example.dtthouses.data.contstant.ErrorMessage.EXCEPTION_ERROR_PREFIX
 
 /**
  * This is a custom exception used to handle network errors
